Rework pyo3 module registration
This allows using `from tket2.circuit import ...`.

drive-by: Remove some commented-out code.
