+++ tket2-py/src/lib.rs
@@ -3,94 +3,36 @@
 
 mod circuit;
 mod optimiser;
-mod pass;
+mod passes;
+mod pattern;
 
-use circuit::{add_circuit_module, to_hugr, T2Circuit};
-use optimiser::add_optimiser_module;
-use pass::add_pass_module;
-
-use hugr::Hugr;
 use pyo3::prelude::*;
-use tket2::portmatching::pyo3::PyPatternMatch;
-use tket2::portmatching::{CircuitPattern, PatternMatcher};
-use tket2::rewrite::CircuitRewrite;
-
-#[derive(Clone)]
-#[pyclass]
-/// A rewrite rule defined by a left hand side and right hand side of an equation.
-pub struct Rule(pub [Hugr; 2]);
-
-#[pymethods]
-impl Rule {
-    #[new]
-    fn new_rule(l: PyObject, r: PyObject) -> PyResult<Rule> {
-        let l = to_hugr(l)?;
-        let r = to_hugr(r)?;
-
-        Ok(Rule([l, r]))
-    }
-}
-#[pyclass]
-struct RuleMatcher {
-    matcher: PatternMatcher,
-    rights: Vec<Hugr>,
-}
-
-#[pymethods]
-impl RuleMatcher {
-    #[new]
-    pub fn from_rules(rules: Vec<Rule>) -> PyResult<Self> {
-        let (lefts, rights): (Vec<_>, Vec<_>) =
-            rules.into_iter().map(|Rule([l, r])| (l, r)).unzip();
-        let patterns: Result<Vec<CircuitPattern>, _> =
-            lefts.iter().map(CircuitPattern::try_from_circuit).collect();
-        let matcher = PatternMatcher::from_patterns(patterns?);
-
-        Ok(Self { matcher, rights })
-    }
-
-    pub fn find_match(&self, target: &T2Circuit) -> PyResult<Option<CircuitRewrite>> {
-        let h = &target.0;
-        let p_match = self.matcher.find_matches_iter(h).next();
-        if let Some(m) = p_match {
-            let py_match = PyPatternMatch::try_from_rust(m, h, &self.matcher)?;
-            let r = self.rights.get(py_match.pattern_id).unwrap().clone();
-            let rw = py_match.to_rewrite(h, r)?;
-            Ok(Some(rw))
-        } else {
-            Ok(None)
-        }
-    }
-}
 
 /// The Python bindings to TKET2.
 #[pymodule]
 #[pyo3(name = "tket2")]
 fn tket2_py(py: Python, m: &PyModule) -> PyResult<()> {
-    add_circuit_module(py, m)?;
-    add_pattern_module(py, m)?;
-    add_pass_module(py, m)?;
-    add_optimiser_module(py, m)?;
+    add_submodule(py, m, circuit::module(py)?)?;
+    add_submodule(py, m, optimiser::module(py)?)?;
+    add_submodule(py, m, passes::module(py)?)?;
+    add_submodule(py, m, pattern::module(py)?)?;
     Ok(())
 }
 
-/// portmatching module
-fn add_pattern_module(py: Python, parent: &PyModule) -> PyResult<()> {
-    let m = PyModule::new(py, "pattern")?;
-    m.add_class::<tket2::portmatching::CircuitPattern>()?;
-    m.add_class::<tket2::portmatching::PatternMatcher>()?;
-    m.add_class::<CircuitRewrite>()?;
-    m.add_class::<Rule>()?;
-    m.add_class::<RuleMatcher>()?;
+fn add_submodule(py: Python, parent: &PyModule, submodule: &PyModule) -> PyResult<()> {
+    parent.add_submodule(submodule)?;
 
-    m.add(
-        "InvalidPatternError",
-        py.get_type::<tket2::portmatching::pattern::PyInvalidPatternError>(),
-    )?;
-    m.add(
-        "InvalidReplacementError",
-        py.get_type::<hugr::hugr::views::sibling_subgraph::PyInvalidReplacementError>(),
-    )?;
-
-    parent.add_submodule(m)
+    // Add submodule to sys.modules.
+    // This is required to be able to do `from parent.submodule import ...`.
+    //
+    // We set both `tket2.tket2.submodule` for the compiled `.so`,
+    // and `tket2.submodule` for the re-exported version.
+    // (See `tket2-py/tket2/__init__.py`).
+    //
+    // See [https://github.com/PyO3/pyo3/issues/759]
+    let submodule_name = submodule.name()?;
+    let modules = py.import("sys")?.getattr("modules")?;
+    modules.set_item(format!("tket2.{submodule_name}"), submodule)?;
+    modules.set_item(format!("tket2.tket2.{submodule_name}"), submodule)?;
+    Ok(())
 }
