--- tket2-py/tket2/custom_base.py
@@ -1,12 +1,0 @@
-# from abc import ABC, abstractmethod
-# from tket2 import RsCircuit, Signature
-
-
-# class CustomOpBase(ABC):
-#     @abstractmethod
-#     def signature(self) -> Signature:
-#         ...
-
-#     @abstractmethod
-#     def to_circuit(self) -> RsCircuit:
-#         ...
